Update the Charon driver to leverage Cargo
