+++ charon/macros/Cargo.toml
@@ -10,3 +10,5 @@ proc-macro = true
 [dependencies]
 syn = "1.0.73"
 quote = "1.0.9"
+serde = { version = "1.0.117", features = ["derive"] }
+toml = "0.5.8"
